feat: 비로그인 상태에서 프로필 접근 막기
