+++ src/main.js
@@ -1,7 +1,8 @@
 import { render } from "./helpers/render.js";
 import { createRouter } from "./helpers/router.js";
-import { routes } from "./routes";
+import { routes, onBeforeEach } from "./router.js";
 
 const root = render("#app");
 
-createRouter(root, routes).start();
+const router = createRouter(root, routes);
+router.beforeEach(onBeforeEach).start();
