+++ src/router.js
@@ -1,0 +1,21 @@
+import { MainPage, NotFoundPage, LoginPage, ProfilePage } from "./pages";
+import { useUserStore } from "./stores/user";
+
+export const routes = {
+  "/": { redirect: "/main" },
+  "/login": { component: LoginPage },
+  "/main": { component: MainPage },
+  "/profile": { component: ProfilePage },
+  "*": { component: NotFoundPage },
+};
+
+export const onBeforeEach = (nextPathname, next) => {
+  const userStore = useUserStore();
+  const isAuthenticated = userStore.userInfo !== null;
+
+  if (!isAuthenticated && nextPathname === "/profile") {
+    next("/login");
+  } else {
+    next();
+  }
+};
