--- src/routes/index.js
@@ -1,9 +1,0 @@
-import { MainPage, NotFoundPage, LoginPage, ProfilePage } from "../pages";
-
-export const routes = {
-  "/": { redirect: "/main" },
-  "/login": { component: LoginPage },
-  "/main": { component: MainPage },
-  "/profile": { component: ProfilePage },
-  "*": { component: NotFoundPage },
-};
